Improve conflict handling
Now there are also handled conflicts for files with different parents, case
which occurs in search.

file_name_changed manages the following case: if there is a file that generates a
conflict, there is needed an extra check, because the conflicting
file may also be in the batch rename.

There was also a problem with navigating between conflicts. If in a search there
were renamed two files with the same name, the arrows would always select the first one.