Change tags handling
Before, there was a variable for each tag, >=0 meaning the position of the tag,
-1 meaning the tag is not currently used and -2 that the tag is unavailable.

This was changed because a variable having multiple meanings is confusing.

Instead of this, now there is used a hashtable that has a structure, so that
for each tag we'll know if it's currently used, if it's available and if it's
used.