Update tag handling
Before, it was checked if a tag is used by using g_strrstr on the entry text.
This function was also used for updating the tag's position and deleting it.

Instead of this, tags are now managed by the insertions/deletions from the entry,
removing g_strrstr entirely from this part of the code.

This was done because of flexibility and reliability reasons.