files-view: don't allow compress if no selection
We weren't checking if there is some file selected to enable/disable
the compress action, which was causing the context menu item be always
displayed if we clicked an element that is not a file.
As odd as it sounds, this happens when allow to expand folders in list
view, since the "empty" element is selectable but not a file.
To fix that, check whether we have a file selected or not.

https://bugzilla.gnome.org/show_bug.cgi?id=771469