files-view: use correct parent when compressing/decompressing
We were using the view as the destination for compression and
decompression.

However, the parent of the files is not always the view. This is the
case in list view when the "allow expand folder" setting is set and the
selection is different than the root location of the view.

To fix it, use a parent of one of the selected files, we assume all
files has the same parent.

https://bugzilla.gnome.org/show_bug.cgi?id=771470